refactor: refactor test data

See merge request 2024ss-se-pr-group/24ss-se-pr-qse-10!39

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/dto/StationDto.java
@@ -3,7 +3,7 @@ package at.ac.tuwien.sepr.groupphase.backend.endpoint.dto;
 public class StationDto {
     private Long id;
     private String name;
-    private Long capacity;
+    private int capacity;
 
     public void setId(Long id) {
         this.id = id;
@@ -13,7 +13,7 @@ public class StationDto {
         this.name = name;
     }
 
-    public void setCapacity(Long capacity) {
+    public void setCapacity(int capacity) {
         this.capacity = capacity;
     }
 
@@ -25,7 +25,7 @@ public class StationDto {
         return name;
     }
 
-    public Long getCapacity() {
+    public int getCapacity() {
         return capacity;
     }
 }
